Project changes & fine tuning on Feature branch

+++ base/api/views.py
@@ -1,31 +1,34 @@
+import requests
 
-from django.utils.timesince import timesince
-from ..models import User, BloodGroup, Profile, Blood_Request, Donation_Criteria_Form
-from django.contrib.auth import authenticate
 from django.urls import reverse
+from django.contrib.auth import authenticate
+from django.utils.timesince import timesince
+from django.contrib.sites.models import Site
 
 from .serializers import *
+from base.utils import create_knox_token
+from ..models import User, BloodGroup, Profile, Blood_Request, Donation_Criteria_Form
+
+
 from knox.auth import AuthToken
 from knox.views import LoginView, LogoutView
-
-from rest_framework.authentication import BasicAuthentication
 from knox.views import LoginView as KnoxLoginView
 
-from rest_framework import status, generics
 from rest_framework.views import APIView
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework import status, generics
 from rest_framework.response import Response
+from rest_framework.authentication import BasicAuthentication
 from rest_framework.permissions import IsAuthenticated, AllowAny
-from base.utils import create_knox_token
+from rest_framework.decorators import api_view, permission_classes
 
 
 from dj_rest_auth.registration.views import SocialLoginView
 from dj_rest_auth.social_serializers import TwitterLoginSerializer
-from allauth.socialaccount.providers.twitter.views import TwitterOAuthAdapter
 from allauth.socialaccount.providers.oauth2.client import OAuth2Client
 from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter
+from allauth.socialaccount.providers.twitter.views import TwitterOAuthAdapter
 from allauth.socialaccount.providers.facebook.views import FacebookOAuth2Adapter
-from ..forms import CreateUserForm
+
 # Create your views here.
 
 
@@ -41,7 +44,7 @@ class LoginAPIView(generics.GenericAPIView):
                 "token": create_knox_token(user=user)[1]
             }, status=status.HTTP_201_CREATED)
         else:
-            return Response({"error": "Invalid credentials"},  status=status.HTTP_400_BAD_REQUEST)
+            return Response({"errors": "Invalid credentials"},  status=status.HTTP_400_BAD_REQUEST)
 
 
 class LogoutAPIView(LogoutView):
@@ -91,6 +94,49 @@ class UserDetailsAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
+class ResetUserPasswordView(APIView):
+    permission_classes = [AllowAny]
+
+    def post(self, request, *args, **kwargs):
+        serializer = PasswordResetSerializer(data=request.data)
+
+        if serializer.is_valid():
+            current_site = Site.objects.get_current()
+            current_site = 'localhost:8000'
+
+            payload = {
+                'uid': kwargs.get('uid'),
+                'token': kwargs.get('token'),
+                'new_password': serializer.validated_data['new_password'],
+                're_new_password': serializer.validated_data['re_new_password']
+            }
+
+            djoser_password_reset_url = 'api/auth/users/reset_password_confirm/'
+
+            protocol = 'https'
+            headers = {'content-Type': 'application/json'}
+
+            if not request.is_secure():
+                protocol = 'http'
+
+            url = f'{protocol}://{current_site}/{djoser_password_reset_url}'
+
+            try:
+                response = requests.post(url, json=payload, headers=headers)
+
+                if response.status_code == 204:
+                    return Response({'message': 'Password Successfully updated'}, status=status.HTTP_200_OK)
+                else:
+                    response_object = response.json()
+
+                    return Response({'errors': response_object}, status=response.status_code)
+
+            except requests.RequestException as e:
+                return Response({'errors': 'Error in making request'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        else:
+            return Response({'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+
+
 # //////////////////////////////////////////////////////////////////
 # SOCIAL AUTH
 # //////////////////////////////////////////////////////////////////
@@ -124,20 +170,6 @@ class DonorHomeAPIView(APIView):
         return Response({'blood_requests': serializer.data}, status=status.HTTP_200_OK)
 
 
-# def donation_comparison_form(user_form_data):
-#     userChoice = []
-#     needed_values = [False, False, False,
-#                      False, False, False, False, False, False, False, True, True, True, False]
-
-#     for field_value in ['rabis_hepatitis_past_1_year', 'tatoo_surgery_past_6_months', 'donated_blood_past_3_months', 'immunisation_past_1_month', 'anitibiotics_past_48_hrs', 'alcoholic_bevarage_past_24_hrs', 'asprin_dentalwork_past_72_hrs', 'cough_common_presently', 'pregnant_breastFeeding_presently', 'menstration_presently', 'health_check_1', 'health_check_2', 'age_betwn_18_60', 'body_weight_less_45']:
-#         choice = user_form_data.cleaned_data.get(field_value)
-#         userChoice.append(choice)
-
-#     for i in range(len(needed_values)):
-#         if needed_values[i] != userChoice[i]:
-#             return False
-#     return True
-
 def donation_comparison_form(user_form_data):
     user_choice = user_form_data.validated_data['qualify']
     return user_choice == True
@@ -168,39 +200,9 @@ class DonationCriteriaAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class LocationMapAPIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Location_map render'}, status=status.HTTP_200_OK)
-
-
-class NotEligibleAPIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Not Eligible'}, status=status.HTTP_200_OK)
-
-
 # //////////////////////////////////////////////////////////////////
 # 3. PATIENTS PAGES
 # //////////////////////////////////////////////////////////////////
-
-class PatientHomeAPIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Patient Home page'}, status=status.HTTP_200_OK)
-
-
-# class ProfileAPIView(APIView):
-#     permission_classes = [IsAuthenticated]
-
-#     def get(self, request, *args, **kwargs):
-#         user_profile = request.user.profile
-#         profile = ProfileFormSerializer(instance=user_profile)
-
-#         return Response({'profile': profile.data}, status=status.HTTP_200_OK)
 
 
 class ProfileAPIView(APIView):
@@ -228,7 +230,7 @@ class ProfileAPIView(APIView):
             userForm.save()
             profileForm.save()
             return Response({'message': 'Profile updated successfully'}, status=status.HTTP_200_OK)
-        return Response({'error_message': 'Please upload a PNG file.'}, status=status.HTTP_400_BAD_REQUEST)
+        return Response({'erros': 'Please upload a PNG file.'}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class RequestBloodAPIView(APIView):
@@ -254,13 +256,6 @@ class RequestBloodAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class RequestSentAPIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Request Sent Successfully'}, status=status.HTTP_200_OK)
-
-
 class PatientHistoryAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
@@ -273,27 +268,22 @@ class PatientHistoryAPIView(APIView):
 
         return Response({'history': serializer.data, 'profileForm': profileForm.data}, status=status.HTTP_200_OK)
 
-    # def get(self, request, *args, **kwargs):
-    #     user = request.user
-    #     user_profile = user.profile
 
-    #     userForm = EditUserFormSerializer(instance=user)
-    #     profileForm = ProfileFormSerializer(instance=user_profile)
-
-    #     return Response({'userForm': userForm.data, 'profileForm': profileForm.data}, status=status.HTTP_200_OK)
-
-
-class DeletePageAPIView(APIView):
+class HospitalAddresses(APIView):
     permission_classes = [IsAuthenticated]
 
-    def get(self, request, pk, *args, **kwargs):
-        item = Blood_Request.objects.get(id=pk)
-        return Response({'item': item}, status=status.HTTP_200_OK)
+    def get(self, request, *args, **kwargs):
+        hospital_address = HospitalAddress.objects.all()
+        hospitals = HospitalAddressSerializer(hospital_address, many=True)
+        return Response({'hospitals': hospitals.data}, status=status.HTTP_200_OK)
 
-    def post(self, request, pk, *args, **kwargs):
-        item = Blood_Request.objects.get(id=pk)
-        item.delete()
-        return Response({'message': 'Deleted successfully'}, status=status.HTTP_200_OK)
+    def post(self, request, *args, **kwargs):
+        serializer = HospitalAddressSerializer(data=request.data)
+        if serializer.is_valid():
+            instance = serializer.save()
+            return Response({'message': 'Hopsital added successfully'}, status=status.HTTP_200_OK)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class DonationAgreement(APIView):
@@ -314,24 +304,14 @@ class DonationAgreement(APIView):
 
         serializer = DonationAgreementSerializer(
             data=request.data, instance=user_profile)
+
         if serializer.is_valid():
             donationAgreement = Donation.objects.create(
                 profile=user_profile,
                 hospital_address=serializer.validated_data['hospital_address'],
+
             )
             return Response({'message': 'Successfully sent'}, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class HospitalAddress(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request, *args, **kwargs):
-        serializer = HospitalAddressSerializer(data=request.data)
-        if serializer.is_valid():
-            instance = serializer.save()
-            return Response({'message': 'Hopsital added successfully'}, status=status.HTTP_200_OK)
-
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -341,13 +321,15 @@ class QuestionsAPIView(APIView):
 
     def get(self, request, *args, **kwargs):
         criteria_quiz = DonationCriteriaQuestions.objects.all()
-        questions = self.serializer_class(
-            criteria_quiz, many=True)
+        questions = self.serializer_class(criteria_quiz, many=True)
         return Response({'questions': questions.data}, status=status.HTTP_200_OK)
 
     def post(self, request, *args, **kwargs):
         user = request.user
         user_profile = user.profile
+
+        if not user.is_staff:
+            return Response({'errors': 'Only admin users can create questions.'}, status=status.HTTP_403_FORBIDDEN)
 
         serializer = self.serializer_class(data=request.data)
 
@@ -360,11 +342,15 @@ class QuestionsAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, *args, **kwargs):
+
+        if not request.user.is_staff:
+            return Response({'errors': 'Only admin users can update questions.'}, status=status.HTTP_403_FORBIDDEN)
+
         question_id = kwargs.get('question_id')
         try:
             question = DonationCriteriaQuestions.objects.get(id=question_id)
         except DonationCriteriaQuestions.DoesNotExist:
-            return Response({'error': 'Question not found'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'errors': 'Question not found'}, status=status.HTTP_404_NOT_FOUND)
 
         serializer = DonationCriteriaQuestionsSerializer(
             question, data=request.data, partial=True)
@@ -375,11 +361,15 @@ class QuestionsAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, *args, **kwargs):
+
+        if not request.user.is_staff:
+            return Response({'errors': 'Only admin users can delete questions.'}, status=status.HTTP_403_FORBIDDEN)
+
         question_id = kwargs.get('question_id')
         try:
             question = DonationCriteriaQuestions.objects.get(id=question_id)
         except DonationCriteriaQuestions.DoesNotExist:
-            return Response({'error': 'Question not found'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'errors': 'Question not found'}, status=status.HTTP_404_NOT_FOUND)
 
         question.delete()
         return Response({'message': 'Question deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
@@ -413,27 +403,3 @@ class NotificationsAPIView(APIView):
                 filtered_requests, many=True)
 
         return Response({'requests': bloodRequests.data}, status=status.HTTP_200_OK)
-
-
-class BloodMatchSuccessAPIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Render blood_match_success.html'}, status=status.HTTP_200_OK)
-
-
-class Error404APIView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, *args, **kwargs):
-        return Response({'message': 'Render error404.html'}, status=status.HTTP_200_OK)
-
-
-# @api_view(["POST"])
-# def userValidateAPI(request):
-
-#     userForm = CreateUserForm(**request.data)
-#     if userForm.is_valid():
-#         return Response({"message": "Form Valid"}, status=status.HTTP_200_OK)
-
-#     return Response(userForm.error_messages, status=status.HTTP_401_UNAUTHORIZED)
